Tokenizer, config, model can be loaded from wandb.
A mixin class is used to share the functionality.
