+++ src/dalle_mini/model/__init__.py
@@ -1,2 +1,3 @@
 from .configuration import DalleBartConfig
 from .modeling import DalleBart
+from .tokenizer import DalleBartTokenizer
